Aggregate versions properly
Code was assuming that version strings are (a) integers only, and (b)
sequential, and thus was only aggregating the version that evaluated as
the highest integer.

Modified the code to look for the version string using a better regex,
not to cast to integer, and to aggregate in order processed.